requirements.txt
torch>=2.0.0
transformers>=4.51.0
numpy>=1.21.0
pathlib>=1.0.1
accelerate>=0.21.0
sentence-transformers>=2.7.0
openai>=1.0.0
python-dotenv>=1.0.0
